requirements.txt
uvicorn
google-generativeai